Refactor FormRenderer into modular components
