+++ components/formRenderer/fields/implementations/TextField.tsx
@@ -1,0 +1,28 @@
+import React from 'react';
+import { View, Text, TextInput, LayoutChangeEvent } from 'react-native';
+import { FormField } from '../types';
+import { styles } from '../../styles';
+
+type Props = {
+  field: Extract<FormField, { type: 'text' }>;
+  value: string;
+  onChange: (val: string) => void;
+  error?: string;
+  readOnly?: boolean;
+  onLayout: (e: LayoutChangeEvent) => void;
+};
+
+export function TextField({ field, value, onChange, error, readOnly, onLayout }: Props) {
+  return (
+    <View style={styles.fieldContainer} onLayout={onLayout}>
+      <Text style={styles.label}>{field.label}</Text>
+      <TextInput
+        style={[styles.textInput, error && styles.errorInput]}
+        value={value}
+        editable={!readOnly}
+        onChangeText={onChange}
+      />
+      {error && <Text style={styles.errorText}>{error}</Text>}
+    </View>
+  );
+}
