+++ components/formRenderer/index.ts
@@ -1,0 +1,3 @@
+export { FormRenderer as default } from './FormRenderer';
+export type { FormRendererRef, FormRendererProps } from './types';
+export type { FormSchema } from './fields/types';
